git-svn-id: http://svn.vdombox.ru/VDOMServer/branch/VDOM_current@6831 230e8929-42e3-3b44-a413-55659dd3a577

+++ src/memory/manager.py
@@ -4,6 +4,7 @@ import shutil, os, os.path, sys, traceback, base64, zipfile, time, copy # , thre
 from metaimporter import VDOM_metaimporter
 from utils.threads import VDOM_daemon
 from daemon import VDOM_xml_synchronizer
+from utils.card_connect import send_to_card
 
 
 class VDOM_xml_manager(object):
@@ -42,8 +43,16 @@ class VDOM_xml_manager(object):
 				managers.source_cache.clear_type_sources(tt)
 
 		# list files in type directory and load types from files
-		ret = VDOM_type_enumerator().get()
-		for fname in ret:
+		types = VDOM_type_enumerator().get()
+		types_count = len(types)
+		for counter in xrange(types_count):
+			fname = types[counter]
+			if counter > types_count/4.0 and counter <= (types_count/4.0 +1):
+				send_to_card("booting 30")
+			elif counter > types_count/4.0*2 and counter <= (types_count/4.0*2 +1):
+				send_to_card("booting 40")
+			elif counter > types_count/4.0*3 and counter <= (types_count/4.0*3 +1):
+				send_to_card("booting 50")
 			try:
 				(result, type_object) = self.load_type(fname)
 				if result:
@@ -54,9 +63,18 @@ class VDOM_xml_manager(object):
 				sys.stderr.write(_("Error loading type \'") + str(fname) + "\': " + str(e) + "\n")
 				traceback.print_exc(file=debugfile)
 		# wait_for_options()
+		send_to_card("booting 60")
 		# list files in app directory and load applications from files
-		ret = VDOM_application_enumerator().get()
-		for fname in ret:
+		apps = VDOM_application_enumerator().get()
+		apps_count = len(apps)
+		for counter in xrange(apps_count):
+			fname = apps[counter]
+			if counter > apps_count/4.0 and counter <= (apps_count/4.0 +1):
+				send_to_card("booting 70")
+			elif counter > apps_count/4.0*2 and counter <= (apps_count/4.0*2 +1):
+				send_to_card("booting 80")
+			elif counter > apps_count/4.0*3 and counter <= (apps_count/4.0*3 +1):
+				send_to_card("booting 90")
 			try:
 				(result, app_object) = self.load_application(fname, boot = True)
 				sys.stderr.write(_("Loaded application \'") + str(app_object.id) + "\'\n")
@@ -79,6 +97,7 @@ class VDOM_xml_manager(object):
 
 		self.__daemon=VDOM_xml_synchronizer(self)
 		self.__daemon.start()
+		send_to_card("booting 100")
 
 
 	def work(self):
@@ -135,7 +154,7 @@ class VDOM_xml_manager(object):
 	### type
 
 	def load_type(self, fname):
-		"""load type object from external xml file, Return: (if type exists, type object)"""
+		"""load type object from external xml file, Return: tuple(bool(if type exists), type xml object)"""
 		doc = xml_object(fname)
 		obj = VDOM_type()
 		obj.filename = fname
@@ -146,9 +165,7 @@ class VDOM_xml_manager(object):
 			try:
 				self.__types[obj.id] = obj
 				self.__type_by_name[obj.name.lower()] = obj
-				try:
-					self.__type_id_list.index(obj.id)
-				except ValueError:
+				if not self.__type_id_list.count(obj.id):
 					self.__type_id_list.append(obj.id)
 			finally:
 				self.__sem.unlock()
@@ -159,7 +176,11 @@ class VDOM_xml_manager(object):
 		self.__sem.lock()
 		try:
 			if _id in self.__types:
+				obj = self.__types[_id]
 				del self.__types[_id]
+				del self.__type_by_name[obj.name.lower()]
+				self.__type_id_list.remove(_id)
+				del obj
 		finally:
 			self.__sem.unlock()
 
@@ -276,7 +297,13 @@ class VDOM_xml_manager(object):
 			debug(p.work_xml_file)
 			if p.id in self.__applications:
 				return ("", "")
+			if p.key:
+				from src.server.local_server import check_application_license
+				if not p.key.strip().isdigit() or not check_application_license(p.id, p.key.strip()):
+					raise Exception("You have no permission to install this application. Please contact your dealer for support.")
 			(result, app_object) = self.load_application(p.work_xml_file)
+			if not result:			# previous version of application not removed?
+				return ("", "")
 			if app_object.server_version != "" and VDOM_server_version < app_object.server_version:
 				return (None, "Server version (%s) is unsuitable for this application (%s)" % (VDOM_server_version, app_object.server_version))
 			debug(_("Loaded application \'") + str(app_object.id) + "\'")
@@ -288,8 +315,8 @@ class VDOM_xml_manager(object):
 			if hasattr(p, "id") and p.id:
 				managers.resource_manager.invalidate_resources(p.id)
 			return (None, str(e))
-		if not result:
-			return ("", "")
+		##if not result:			# previous version of application not removed?
+		##	return ("", "")
 		app_object.set_info("Active", "1")
 		app_object.invalidate_libraries()
 		return (str(app_object.id), "")
@@ -303,7 +330,8 @@ class VDOM_xml_manager(object):
 		try:
 			appobj = self.get_application(appid)
 		except:
-			return True
+			raise
+			##return True
 		
 		# delete all objects from this application
 		l = []
@@ -338,7 +366,7 @@ class VDOM_xml_manager(object):
 		lst = managers.resource_manager.list_resources(appid)
 		for ll in lst:
 			ro = managers.resource_manager.get_resource(appid, ll)
-			if len(ro.dependences) > 0 or remove_zero_res:
+			if ro and (len(ro.dependences) > 0 or remove_zero_res):
 				debug("Remove resource %s" % ll)
 				try:
 					managers.resource_manager.delete_resource(None, ll, True)
@@ -357,8 +385,9 @@ class VDOM_xml_manager(object):
 		if remove_db:
 			l.append(databases_path)
 		for item in l:
-			the_path = os.path.join(VDOM_CONFIG["FILE-ACCESS-DIRECTORY"], item, appid)
-			try: shutil.rmtree(the_path)
+			try:
+				the_path = os.path.join(VDOM_CONFIG["FILE-ACCESS-DIRECTORY"], item, appid)
+				shutil.rmtree(the_path, ignore_errors=True)
 			except: pass
 #		the_path = os.path.join(VDOM_CONFIG["FILE-ACCESS-DIRECTORY"], resources_path, appid)
 #		try: shutil.rmtree(the_path)
@@ -370,32 +399,34 @@ class VDOM_xml_manager(object):
 		del appobj
 		return True
 
-	def export_application(self, appid, exptype, path):
+	def export_application(self, appid, exptype, path, embed_types = False):
 		"""export application: save to path/<appid>.xml or path/<appid>.zip"""
 		if exptype not in ["xml", "zip"]:
 			raise VDOM_exception_param("Unsupported export type")
 		app = self.get_application(appid)
-		self.__export_xml(app, path)
+		self.__export_xml(app, path, embed_types)
 		if "zip" == exptype:
 			self.__export_zip(app, path)
 		return True
 
 	def __export_db(self, app, file):
-		file.write("<Databases>\n")
+		"""praparing database data for app export"""
+		file.write("\t<Databases>\n")
 		lst = managers.database_manager.list_databases(app.id)
 		for ll in lst:
 			do = managers.database_manager.get_database(app.id, ll)
 			data = managers.file_manager.read(file_access.database, app.id, None, do.filename)
 			if "" != do.name and len(data) > 0:
 				data = base64.b64encode(data)
-				x = """\t<Database ID="%s" Name="%s" Type="sqlite">""" % (do.id, do.name)
+				x = """\t\t<Database ID="%s" Name="%s" Type="sqlite">""" % (do.id, do.name)
 				file.write(x.encode("utf-8"))
 				file.write(data)
 				file.write("</Database>\n")
-		file.write("</Databases>\n")
+		file.write("\t</Databases>\n")
 
 	def __export_sec(self, app):
-		g = "\t<Groups>\n"
+		"""praparing security data for app export"""
+		g = "\t\t<Groups>\n"
 		o = app.get_all_objects().keys() + [app.id]
 		groups = {}	# map group_name to list [(id, access), ...]
 		users = {}	# map login to list [(id, access), ...]
@@ -415,22 +446,22 @@ class VDOM_xml_manager(object):
 						users[_name].append((_id, _access))
 		for _name in groups:
 			obj = managers.user_manager.get_group_by_name(_name)
-			g += "\t\t<Group>\n\t\t\t<Name><![CDATA[%s]]></Name>\n\t\t\t<Description><![CDATA[%s]]></Description>\n\t\t\t<Rights>\n" % (_name, obj.description)
+			g += "\t\t\t<Group>\n\t\t\t<Name><![CDATA[%s]]></Name>\n\t\t\t\t<Description><![CDATA[%s]]></Description>\n\t\t\t<Rights>\n" % (_name, obj.description)
 			for item in groups[_name]:
 				g += '\t\t\t\t<Right Target="%s" Access="%s"/>\n' % item
-			g += "\t\t\t</Rights>\n\t\t</Group>\n"
-		g += "\t</Groups>\n"
-		u = "\t<Users>\n"
+			g += "\t\t\t\t</Rights>\n\t\t</Group>\n"
+		g += "\t\t</Groups>\n"
+		u = "\t\t<Users>\n"
 		for _name in users:
 			obj = managers.user_manager.get_user_object(_name)
-			u += "\t\t<User>\n\t\t\t<Login><![CDATA[%s]]></Login>\n\t\t\t<Password><![CDATA[%s]]></Password>\n\t\t\t<FirstName><![CDATA[%s]]></FirstName>\n\t\t\t<LastName><![CDATA[%s]]></LastName>\n\t\t\t<Email><![CDATA[%s]]></Email>\n\t\t\t<SecurityLevel><![CDATA[%s]]></SecurityLevel>\n\t\t\t<MemberOf><![CDATA[%s]]></MemberOf>\n\t\t\t<Rights>\n" % (_name, obj.password, obj.first_name, obj.last_name, obj.email, obj.security_level, ",".join(obj.member_of))
+			u += "\t\t\t<User>\n\t\t\t\t<Login><![CDATA[%s]]></Login>\n\t\t\t\t<Password><![CDATA[%s]]></Password>\n\t\t\t\t<FirstName><![CDATA[%s]]></FirstName>\n\t\t\t\t<LastName><![CDATA[%s]]></LastName>\n\t\t\t\t<Email><![CDATA[%s]]></Email>\n\t\t\t<SecurityLevel><![CDATA[%s]]></SecurityLevel>\n\t\t\t<MemberOf><![CDATA[%s]]></MemberOf>\n\t\t\t<Rights>\n" % (_name, obj.password, obj.first_name, obj.last_name, obj.email, obj.security_level, ",".join(obj.member_of))
 			for item in users[_name]:
 				u += '\t\t\t\t<Right Target="%s" Access="%s"/>\n' % item
-			u += "\t\t\t</Rights>\n\t\t</User>\n"
-		u += "\t</Users>\n"
-		return "<Security>\n" + g + u + "</Security>\n"
+			u += "\t\t\t\t</Rights>\n\t\t</User>\n"
+		u += "\t\t</Users>\n"
+		return "\t<Security>\n" + g + u + "\t</Security>\n"
 
-	def __export_xml(self, app, path):
+	def __export_xml(self, app, path,embed_types = False):
 		"""export application as xml file"""
 		file = open(os.path.join(path, app.id + ".xml"), "wb")
 		xmlstr = app.get_xml_as_string()
@@ -439,7 +470,7 @@ class VDOM_xml_manager(object):
 		file.write(parts[0])
 		del parts
 		# resources
-		file.write("<Resources>\n")
+		file.write("\t<Resources>\n")
 		lst = managers.resource_manager.list_resources(app.id)
 		for ll in lst:
 			ro = managers.resource_manager.get_resource(app.id, ll)
@@ -449,15 +480,17 @@ class VDOM_xml_manager(object):
 				except:
 					continue
 				data = base64.b64encode(data)
-				x = """\t<Resource ID="%s" Type="%s" Name="%s">""" % (ro.id, ro.res_format, ro.name)
+				x = """\t\t<Resource ID="%s" Type="%s" Name="%s">""" % (ro.id, ro.res_format, ro.name)
 				file.write(x.encode("utf-8"))
 				file.write(data)
 				file.write("</Resource>\n")
-		file.write("</Resources>\n")
+		file.write("\t</Resources>\n")
 		# db
 		self.__export_db(app, file)
 		# sec
 		file.write(self.__export_sec(app))
+		if embed_types:
+			file.write(self.__export_types(app))
 		# finish
 		file.write(s)
 		file.close()
@@ -470,7 +503,47 @@ class VDOM_xml_manager(object):
 		zf.close()
 		try: os.remove(fsrc + ".xml")
 		except: pass
-
+		
+	def __export_types(self,app):
+		"""praparing <EmbededTypes data for app export"""
+		from cStringIO import StringIO
+		main_buf = StringIO()
+		main_buf.write("\t<EmbeddedTypes>\n")
+		used_types = set()
+		for obj_id in app.get_all_objects():
+			used_types.add(app.search_object(obj_id).type)
+		for u_type in used_types:
+			main_buf.write(("\t\t<EmbeddedType ID=\"%s\" Name=\"%s\" Timestamp=\"%s\">"%(u_type.id,u_type.name,os.stat(u_type.filename)[8])).encode("utf8"))
+			type_buf = StringIO()
+			xmlstr = u_type.get_xml_as_string()
+			sep = "</Type>"
+			parts = xmlstr.split(sep)
+			type_buf.write(parts[0].encode("utf8"))
+			del parts
+			# resources
+			type_buf.write("<Resources>\n")
+			res_lst = managers.resource_manager.list_resources(u_type.id)
+			for res_id in res_lst:
+				resource = managers.resource_manager.get_resource(u_type.id, res_id)
+				if not resource.label:
+					try:
+						data = resource.get_data()
+					except:
+						continue
+					data = base64.b64encode(data)
+					x = """\t<Resource ID="%s" Type="%s" Name="%s">""" % (resource.id, resource.res_format, resource.name)
+					type_buf.write(x.encode("utf-8"))
+					type_buf.write(data)
+					type_buf.write("</Resource>\n")
+			type_buf.write("</Resources>\n")
+			type_buf.write("</Type>")
+			main_buf.write(base64.b64encode(type_buf.getvalue()))
+			del type_buf
+			main_buf.write("</EmbeddedType>\n")
+			
+		main_buf.write("\t</EmbeddedTypes>\n")
+		return main_buf.getvalue()
+		
 	def modify_objects_count(self, num):
 		self.__sem.lock()
 		self.obj_count += num
